完善 Sprite

+++ Source/Clem/Rendering/Sprite.cpp
@@ -36,13 +36,13 @@ void Sprite::drawPoint(const Point2i& p, const Tile& t)
 
 void Sprite::drawLine(const Point2i& start, const Point2i& end, const Tile& t)
 {
-	const float dx  = end.x - start.x;
-	const float dy  = end.y - start.y;
+	const float dx  = (float)end.x - start.x;
+	const float dy  = (float)end.y - start.y;
 	const float max = std::max(dx, dy);
 	const float kx  = max / dx;
 	const float ky  = max / dy;
 
-	float x = start.x, y = start.y;
+	float x = (float)start.x, y = (float)start.y;
 	for(int i = 0; i <= max; i++)
 	{
 		drawPoint((int)(x + 0.5f), (int)(y + 0.5f), t);
